+++ qt_ui/main.py
@@ -353,7 +353,6 @@ def create_game(
             f100_supersabre=False,
             f104_starfighter=False,
             f105_thunderchief=False,
-            hercules=False,
             oh_6=False,
             oh_6_vietnamassetpack=False,
             uh_60l=False,